Add the ability to group functions and execute group tests.
I found myself needing to run only a group of tests, but at different
times. Picking a common set of tests out and appending 'focus_' to their names
was not entirely workable--it took time to do this and was error prone. What I
needed was the ability to run either all tests or a single group.

That's what this commit allows.

Note that etest-runner now parses command line arguments, the focus concept has
been dropped (as you can create a focus group) and the internal bits of
etest_runner.erl have been shuffled to accommodate the group concept. Note also
that my_second_test.erl has been updated to showcase the group concept. Try it
out with

    bin/etest-runner -g slow

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>